+++ pkg/analysis_server/test/stress/replay/replay.dart
@@ -3,7 +3,6 @@
 // BSD-style license that can be found in the LICENSE file.
 
 /// A stress test for the analysis server.
-import 'dart:async';
 import 'dart:io';
 import 'dart:math' as math;