+++ pkg/analyzer_cli/tool/perf.dart
@@ -3,7 +3,6 @@
 // BSD-style license that can be found in the LICENSE file.
 
 /// An entrypoint used to run portions of analyzer and measure its performance.
-import 'dart:async';
 import 'dart:io' show exit;
 
 import 'package:analyzer/dart/analysis/features.dart';